requirements.txt
langchain
langchain-openai
textwrap3
python-dotenv
fastapi
uvicorn